Created proper folder structure for Matlab bindings
Also moved the Eigen Matrix to Matlab conversion to a separate file
